refactor: reorganize project structure and module hierarchy
