--- aiaccel/hpo/apps/optimize/__main__.py
@@ -1,4 +1,0 @@
-from aiaccel.hpo.apps.optimize.optimize import main
-
-if __name__ == "__main__":
-    main()
