+++ aiaccel/hpo/job_executors/local_job_executor.py
@@ -8,8 +8,8 @@ import subprocess
 import time
 import traceback
 
-from aiaccel.hpo.job.executors.base_job_executor import BaseJobExecutor
-from aiaccel.hpo.job.jobs.local_job import LocalJob
+from aiaccel.hpo.job_executors.base_job_executor import BaseJobExecutor
+from aiaccel.hpo.jobs.local_job import LocalJob
 
 
 class LocalJobExecutor(BaseJobExecutor):
