+++ aiaccel/torch/apps/config/train_ddp.yaml
