--- aiaccel/torch/apps/train/__main__.py
@@ -1,4 +1,0 @@
-from aiaccel.torch.apps.train.train import main
-
-if __name__ == "__main__":
-    main()
