+++ dependencies.yaml
@@ -1,0 +1,48 @@
+optional-dependencies:
+  # Development dependencies
+  dev:
+    - mypy
+    - myst-parser
+    - pre-commit
+    - pytest
+    - pytest-cov
+    - pytest-mock
+    - pytest-subprocess
+    - ruff
+    - sphinx
+    - sphinxcontrib-jquery
+    - sphinx-intl
+    - sphinx-fontawesome
+    - sphinx-rtd-theme
+    - types-colorama
+    - undecorated
+    - pandas
+    - pandas-stubs
+    - matplotlib
+    - optuna>=3.4.0
+    - torch>=2.2.0
+    - lightning>=2.2.1
+    - h5py
+
+  # Hyperparameter optimization dependencies
+  hpo:
+    - optuna>=3.4.0
+
+  # PyTorch related dependencies
+  torch:
+    - torch>=2.2.0
+    - lightning>=2.2.1
+    - h5py
+    - pandas
+    - pandas-stubs
+    - matplotlib
+
+  # Full feature set
+  all:
+    - optuna>=3.4.0
+    - torch>=2.2.0
+    - lightning>=2.2.1
+    - h5py
+    - pandas
+    - pandas-stubs
+    - matplotlib
